requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
pyjwt
pwdlib[argon2]
loguru
python-multipart
